Refactor fine collection service
