+++ python/FineCollectionService/fine_collection/clients.py
@@ -2,6 +2,11 @@ import requests
 from pydantic import BaseModel
 
 
+class ClientError(Exception):
+    def  __init__(self, message: str):
+        super().__init__(message)
+
+
 class Vehicle(BaseModel):
     vehicle_id: str
     make: str
@@ -14,8 +19,10 @@ class VehicleRegistrationClient:
     def __init__(self, base_address: str):
         self.base_address = base_address
 
-    def get_vehicle_info(self, license_number):
+    def get_vehicle_info(self, license_number: str) -> Vehicle:
         response = requests.get(f"{self.base_address}/vehicleinfo/{license_number}")
-        response_data = Vehicle.parse_raw(response.content)
 
-        return response_data
+        if response.status_code == 200:
+            return Vehicle.parse_raw(response.content)
+        
+        raise ClientError("Failed to retrieve vehicle data. Got status %i: %s", response.status_code, str(response.content))
